added api for withdrawal transactions

+++ backend/BankingApp/src/main/java/com/bankingapp/repository/AccountRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bankingapp.models.Account;
 
 @Repository
-public interface AccountRepo extends JpaRepository<Account, Integer> {
+public interface AccountRepo extends JpaRepository<Account, Long> {
 	
 }
